steiner: apply §8 cosmetic fixes from review
- Fix steiner_flags_uint8 type: std::vector<int> → std::vector<uint8_t>
  (4× memory saving for large point clouds)
- Remove unpopulated m_steiner_graph_terminal_indices member from
  SteinerGrapher.h; leave note for future recover_steiner_graph() port
- Remove dead commented-out debug/cout blocks from CreateSteinerGraph.cxx

+++ clus/src/SteinerGrapher.cxx
@@ -1027,7 +1027,7 @@ EnhancedSteinerResult create_enhanced_steiner_graph(
 
         // Add the flag_steiner_terminal boolean array to point cloud
         // Convert std::vector<bool> to std::vector<uint8_t> to avoid std::vector<bool> specialization issues
-        std::vector<int> steiner_flags_uint8(result.flag_steiner_terminal.begin(), result.flag_steiner_terminal.end());
+        std::vector<uint8_t> steiner_flags_uint8(result.flag_steiner_terminal.begin(), result.flag_steiner_terminal.end());
         PointCloud::Array steiner_flag_array(steiner_flags_uint8);
         result.point_cloud.add("flag_steiner_terminal", std::move(steiner_flag_array));
     }
